+++ app/modbus.mjs
@@ -10,6 +10,7 @@ const AVAILABLE_FLAGS = {
     'maxCooling': 7,
     'manualBoost': 10,
     'summerNightCooling': 12,
+    'eco': 40,
 }
 
 // Modes that can only be true one at a time (mapped to their coil number)
@@ -20,6 +21,7 @@ const MUTUALLY_EXCLUSIVE_MODES = {
     'maxHeating': 6,
     'maxCooling': 7,
     'manualBoost': 10,
+    'eco': 40,
 }
 
 const AVAILABLE_SETTINGS = {
@@ -81,6 +83,12 @@ export const getFlagSummary = async (modbusClient) => {
         'maxCooling': result.data[7],
         'manualBoost': result.data[10],
         'summerNightCooling': result.data[12],
+    }
+
+    result = await mutex.runExclusive(async () => modbusClient.readCoils(40, 1))
+    summary = {
+        ...summary,
+        'eco': result.data[0],
     }
 
     return summary
@@ -249,7 +257,7 @@ export const getDeviceInformation = async (modbusClient) => {
     result = await mutex.runExclusive(async () => modbusClient.readHoldingRegisters(171, 1))
     deviceInformation = {
         ...deviceInformation,
-        'heatingTypeInstalled': getHeatingTypeName(result.data[0]),
+        'heatingTypeInstalled': getAutomationAndHeatingTypeName(result.data[0]),
     }
 
     result = await mutex.runExclusive(async () => modbusClient.readHoldingRegisters(597, 3))
@@ -382,22 +390,22 @@ const getCoolingTypeName = (coolingTypeInt) => {
     ][coolingTypeInt]
 }
 
-export const getHeatingTypeName = (heatingTypeInt) => {
-    // 0=Ei lämmitintä, 1=VPK, 2=HP, 3=SLP, 4=SLP PWM. Mapping known values to the actual names used on the product,
-    // these seem to be internal
+export const getAutomationAndHeatingTypeName = (heatingTypeInt) => {
+    // 0=Ei lämmitintä, 1=VPK, 2=HP, 3=SLP, 4=SLP PWM
+    // E prefix is used for units with EDA automation
+    // M prefix is used for units with MD automation
     return (
         [
-            'ED', // prettier-hack
-            'EDW',
-            'HP',
-            'EDE',
-            'SLP PWM',
+            'ED/MD', // prettier-hack
+            'EDW/MDW',
+            'EDX/MDX',
+            'EDE/MDE',
         ][heatingTypeInt] || 'unknown'
     )
 }
 
 export const createModelNameString = (deviceInformation) => {
-    // E.g. LTR-3 eco EDE - CG
+    // E.g. LTR-3 eco EDE/MDE - CG
     let modelName = deviceInformation.familyType
 
     if (deviceInformation.fanType === 'EC') {